Create CoCo1and2RayTrace.txt
Color Computer 1 & 2 version of the BASIC ray tracing code.
Includes the Hitachi 6309 control code, but leaves out the the math library patch.  I could copy the ROM to RAM and patch it, but then the CPU high speed clock enable wouldn't work.

Uses NTSC color artifacting in the 2 color mode to display red and blue colors. 

The MC6847 color artifacting may be reversed, and the machine needs to be reset until the colors come up correctly.  There is no way to control this from a program, it's caused by the state the 6847 comes up in on reset.

The image will be black & white on PAL systems, and different colors on clone computers from South America.